refactor: consolidate files in 'src/auth/credentials'
An existing file had a small amount of code and
could instead be moved to the utils.ts file.
