+++ src/auth/providers/sharedCredentialsProvider.ts
@@ -12,8 +12,7 @@ import { fromInstanceMetadata, fromContainerMetadata } from '@aws-sdk/credential
 import { fromEnv } from '@aws-sdk/credential-provider-env'
 import { getLogger } from '../../shared/logger'
 import { getStringHash } from '../../shared/utilities/textUtilities'
-import { getMfaTokenFromUser } from '../credentials/credentialsCreator'
-import { resolveProviderWithCancel } from '../credentials/utils'
+import { getMfaTokenFromUser, resolveProviderWithCancel } from '../credentials/utils'
 import { CredentialsProvider, CredentialsProviderType, CredentialsId } from './credentials'
 import { CredentialType } from '../../shared/telemetry/telemetry.gen'
 import { assertHasProps, getMissingProps, hasProps } from '../../shared/utilities/tsUtils'
